+++ tests/functional_tests/CMakeLists.txt
@@ -66,7 +66,7 @@ target_sources(
     workgroup_primitives.cpp
     empty_tester.cpp
     wavefront_primitives.cpp
-    put_a2a_tester.cpp
+    flood_tester.cpp
 )
 
 ###############################################################################
